example/docker/spva_jwt/requirements.txt
flask==3.0.3
PyJWT==2.9.0
cryptography==43.0.1
